fix: 회원가입 절차 변경에 따른 ui 반영
